SPIM syscalls are available to all programs. Custom syscalls are added using the USER_SYSCALL() macro.

+++ hostapps/load-lvl1.c
@@ -44,7 +44,7 @@
 int main(int argc, char **argv)
 {
 	char *base, *l1elf, *l2elf;
-	size_t l1sz, l2sz;
+	unsigned int l1sz, l2sz;
 	Elf32_Sym *sl2_elf, *sl2_elf_size;
 	struct mips_cpu *pcpu;
